Refactor AgentManager: extract focused managers
Split the 1,435-line AgentManager god object into focused single-responsibility managers:

- WorktreeManager: git worktree CRUD, coordination file copying, scanning
- AgentStatusTracker: status polling, icon updates, approval tracking
- AgentPersistence: save/restore agents, VS Code state + worktree metadata

AgentManager is now a thin coordinator (~650 lines) that delegates to specialized managers.
