+++ vscode-extension/src/agentManager.ts
@@ -3,8 +3,8 @@ import * as path from 'path';
 import * as fs from 'fs';
 import { agentPath } from './pathUtils';
 import { ContainerManager } from './containerManager';
+import { WorktreeManager, AgentStatusTracker, AgentPersistence } from './managers';
 
-// Import types from centralized types module
 import {
     Agent,
     AgentStatus,
@@ -14,510 +14,106 @@ import {
     IsolationTier,
     ContainerInfo,
     AGENT_NAMES,
-    STATUS_ICONS,
 } from './types';
 
-// Import services
 import {
     getConfigService,
     getCommandService,
     getGitService,
     getTerminalService,
-    getStatusService,
     getEventBus,
     getLogger,
     isLoggerInitialized,
     getTerminalIcon,
-    getPersistenceService,
-    isPersistenceServiceInitialized,
 } from './services';
 
 // Re-export types for backward compatibility
 export { Agent, AgentStatus, PersistedAgent, DiffStats, PendingApproval, IsolationTier, ContainerInfo };
 
+/**
+ * Coordinates agent lifecycle and delegates to specialized managers.
+ */
 export class AgentManager {
     private agents: Map<number, Agent> = new Map();
     private workspaceRoot: string;
     private extensionPath: string;
+
+    // Specialized managers
     private containerManager: ContainerManager;
+    private worktreeManager: WorktreeManager;
+    private statusTracker: AgentStatusTracker;
+    private persistence: AgentPersistence;
 
     constructor(extensionPath: string) {
         this.workspaceRoot = vscode.workspace.workspaceFolders?.[0]?.uri.fsPath || '';
         this.extensionPath = extensionPath;
+
+        // Initialize managers
         this.containerManager = new ContainerManager(extensionPath);
+        this.worktreeManager = new WorktreeManager(extensionPath);
+        this.statusTracker = new AgentStatusTracker();
+        this.persistence = new AgentPersistence(this.worktreeManager, this.containerManager);
     }
 
-    // Convenience getter for worktree directory
     private get worktreeDir(): string {
         return getConfigService().worktreeDirectory;
     }
 
-    // Must be called after construction to enable persistence
     setContext(context: vscode.ExtensionContext): void {
         this.containerManager.setContext(context);
-        this.restoreAgents();
+        this.agents = this.persistence.restoreAgents(this.getRepositoryPaths());
     }
 
-    // Get the container manager for external access
     getContainerManager(): ContainerManager {
         return this.containerManager;
     }
 
-    // Get available isolation tiers from container manager
     async getAvailableIsolationTiers(): Promise<IsolationTier[]> {
         return this.containerManager.getAvailableTiers();
     }
 
-    // Change isolation tier for an existing agent
-    async changeAgentIsolationTier(agentId: number, newTier: IsolationTier): Promise<boolean> {
-        const agent = this.agents.get(agentId);
-        if (!agent) {
-            this.debugLog(`Cannot change tier: agent ${agentId} not found`);
-            return false;
-        }
-
-        const currentTier = agent.isolationTier || 'standard';
-        if (currentTier === newTier) {
-            this.debugLog(`Agent ${agentId} already at tier ${newTier}`);
-            return true;
-        }
-
-        this.debugLog(`Changing agent ${agentId} isolation from ${currentTier} to ${newTier}`);
-
-        try {
-            // Remove existing container if containerized
-            if (currentTier !== 'standard') {
-                await this.containerManager.removeContainer(agentId);
-            }
-
-            // Create new container if not going to standard
-            if (newTier !== 'standard') {
-                const repoConfig = this.containerManager.loadRepoConfig(agent.repoPath);
-                const containerInfo = await this.containerManager.createContainer(
-                    agentId,
-                    agent.worktreePath,
-                    newTier,
-                    repoConfig
-                );
-                agent.containerInfo = containerInfo;
-            } else {
-                agent.containerInfo = undefined;
-            }
-
-            agent.isolationTier = newTier;
-            this.saveAgents();
-
-            return true;
-        } catch (error) {
-            this.debugLog(`Failed to change isolation tier: ${error}`);
-            vscode.window.showErrorMessage(`Failed to change isolation tier: ${error}`);
-            return false;
-        }
-    }
-
-    // Get container stats for an agent
-    async getAgentContainerStats(agentId: number): Promise<{ memoryMB: number; cpuPercent: number } | null> {
-        return this.containerManager.getContainerStats(agentId);
-    }
-
-    // Debug logging via Logger service
     private debugLog(message: string): void {
         if (isLoggerInitialized()) {
             getLogger().child('AgentManager').debug(message);
         }
     }
 
-    // Generate a UUID for Claude session
-    private generateSessionId(): string {
-        return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, (c) => {
-            const r = Math.random() * 16 | 0;
-            const v = c === 'x' ? r : (r & 0x3 | 0x8);
-            return v.toString(16);
-        });
-    }
-
-    // Get agent name from NATO phonetic alphabet
     private getAgentName(id: number): string {
         if (id <= AGENT_NAMES.length) {
             return AGENT_NAMES[id - 1];
         }
-        // For IDs beyond the list, use name + number (e.g., alpha-2)
         const baseIndex = (id - 1) % AGENT_NAMES.length;
         const suffix = Math.floor((id - 1) / AGENT_NAMES.length) + 1;
         return `${AGENT_NAMES[baseIndex]}-${suffix}`;
     }
 
-    // Save agents to persistent storage (both VS Code state and worktree metadata)
-    private saveAgents(): void {
-        // Save to VS Code workspace state
-        if (isPersistenceServiceInitialized()) {
-            getPersistenceService().saveAgents(this.agents);
-        }
-
-        // Save to worktree metadata files (source of truth)
-        for (const agent of this.agents.values()) {
-            this.saveAgentToWorktree(agent);
-        }
-    }
-
-    // Restore agents from worktree metadata (primary) and VS Code state (fallback)
-    private restoreAgents(): void {
-        this.debugLog(`[restoreAgents] Starting agent restoration`);
-
-        // First, scan worktrees for agents (this is the source of truth)
-        const repoPaths = this.getRepositoryPaths();
-        const worktreeAgents = new Map<string, PersistedAgent>();
-
-        for (const repoPath of repoPaths) {
-            const agents = this.scanWorktreesForAgents(repoPath);
-            for (const agent of agents) {
-                // Use worktreePath as unique key
-                worktreeAgents.set(agent.worktreePath, agent);
-            }
-        }
-
-        this.debugLog(`[restoreAgents] Found ${worktreeAgents.size} agents in worktrees`);
-
-        // Also load from VS Code state as fallback (for agents without worktree metadata)
-        let vscodeAgents: PersistedAgent[] = [];
-        if (isPersistenceServiceInitialized()) {
-            vscodeAgents = getPersistenceService().loadPersistedAgents();
-            this.debugLog(`[restoreAgents] Found ${vscodeAgents.length} agents in VS Code state`);
-        }
-
-        // Merge: worktree metadata takes priority
-        const allAgents = new Map<string, PersistedAgent>();
-
-        // Add VS Code state agents first
-        for (const agent of vscodeAgents) {
-            allAgents.set(agent.worktreePath, agent);
-        }
-
-        // Override with worktree metadata (source of truth)
-        for (const [path, agent] of worktreeAgents) {
-            allAgents.set(path, agent);
-        }
-
-        // Log available terminals for debugging
-        const terminalNames = vscode.window.terminals.map(t => t.name);
-        this.debugLog(`[restoreAgents] Available terminals: ${JSON.stringify(terminalNames)}`);
-
-        // Create Agent objects from persisted data
-        for (const persisted of allAgents.values()) {
-            // Try to find existing terminal for this agent
-            const existingTerminal = vscode.window.terminals.find(
-                t => t.name === persisted.name
-            );
-
-            // Get container info if this agent is containerized
-            const containerInfo = this.containerManager.getContainer(persisted.id);
-
-            const agent: Agent = {
-                ...persisted,
-                // Generate sessionId for old agents that don't have one
-                sessionId: persisted.sessionId || this.generateSessionId(),
-                terminal: existingTerminal || null,
-                status: 'idle',
-                statusIcon: existingTerminal ? 'circle-filled' : 'circle-outline',
-                pendingApproval: null,
-                lastInteractionTime: new Date(),
-                diffStats: { insertions: 0, deletions: 0, filesChanged: 0 },
-                containerInfo,
-            };
-
-            this.agents.set(agent.id, agent);
-            this.debugLog(`[restoreAgents] Restored agent ${agent.name} (id=${agent.id})`);
-        }
-
-        this.debugLog(`[restoreAgents] Restored ${this.agents.size} total agents`);
-    }
-
-    // Convert a path to the format expected by the configured terminal
     private toTerminalPath(inputPath: string): string {
         return agentPath(inputPath).forTerminal();
     }
 
-    // Convert any path format to Windows path (for VS Code APIs like terminal cwd)
     private toWindowsPath(inputPath: string): string {
         return agentPath(inputPath).forNodeFs();
     }
 
-    // Execute a command using CommandService
     private execCommand(command: string, cwd: string): string {
         return getCommandService().exec(command, cwd);
     }
 
-    // Execute a command silently using CommandService
     private execCommandSilent(command: string, cwd: string): void {
         getCommandService().execSilent(command, cwd);
     }
 
-    // Recursively copy a directory
-    private copyDirRecursive(src: string, dest: string): void {
-        fs.mkdirSync(dest, { recursive: true });
-        const entries = fs.readdirSync(src, { withFileTypes: true });
-
-        for (const entry of entries) {
-            const srcPath = path.join(src, entry.name);
-            const destPath = path.join(dest, entry.name);
-
-            if (entry.isDirectory()) {
-                this.copyDirRecursive(srcPath, destPath);
-            } else {
-                fs.copyFileSync(srcPath, destPath);
-            }
-        }
-    }
-
     getRepositoryPaths(): string[] {
         const configuredPaths = getConfigService().repositoryPaths;
-
         if (configuredPaths.length > 0) {
             return configuredPaths;
         }
-
-        // Fall back to workspace root
         return this.workspaceRoot ? [this.workspaceRoot] : [];
     }
 
     // ========================================================================
-    // Worktree Metadata Persistence
+    // Agent Lifecycle
     // ========================================================================
-
-    private readonly METADATA_DIR = '.opus-orchestra';
-    private readonly METADATA_FILE = 'agent.json';
-
-    /**
-     * Save agent metadata to the worktree for persistence across sessions
-     */
-    private saveAgentToWorktree(agent: Agent): void {
-        try {
-            const worktreePath = agentPath(agent.worktreePath);
-            const metadataDir = worktreePath.join(this.METADATA_DIR).forNodeFs();
-            const metadataFile = `${metadataDir}/${this.METADATA_FILE}`;
-
-            fs.mkdirSync(metadataDir, { recursive: true });
-
-            const metadata: PersistedAgent = {
-                id: agent.id,
-                name: agent.name,
-                sessionId: agent.sessionId,
-                branch: agent.branch,
-                worktreePath: agent.worktreePath,
-                repoPath: agent.repoPath,
-                taskFile: agent.taskFile,
-                isolationTier: agent.isolationTier,
-                sessionStarted: agent.sessionStarted,
-            };
-
-            fs.writeFileSync(metadataFile, JSON.stringify(metadata, null, 2));
-            this.debugLog(`Saved agent metadata to ${metadataFile}`);
-        } catch (error) {
-            this.debugLog(`Failed to save agent metadata: ${error}`);
-        }
-    }
-
-    /**
-     * Load agent metadata from a worktree
-     */
-    private loadAgentFromWorktree(worktreePath: string): PersistedAgent | null {
-        try {
-            const wtPath = agentPath(worktreePath);
-            const metadataFile = wtPath.join(this.METADATA_DIR, this.METADATA_FILE).forNodeFs();
-
-            if (!fs.existsSync(metadataFile)) {
-                return null;
-            }
-
-            const content = fs.readFileSync(metadataFile, 'utf-8');
-            const metadata = JSON.parse(content) as PersistedAgent;
-            this.debugLog(`Loaded agent metadata from ${metadataFile}`);
-            return metadata;
-        } catch (error) {
-            this.debugLog(`Failed to load agent metadata from ${worktreePath}: ${error}`);
-            return null;
-        }
-    }
-
-    /**
-     * Scan worktrees directory for existing agents and restore them
-     */
-    private scanWorktreesForAgents(repoPath: string): PersistedAgent[] {
-        const agents: PersistedAgent[] = [];
-        const repoTerminalPath = this.toTerminalPath(repoPath);
-        const worktreesDir = agentPath(`${repoTerminalPath}/${this.worktreeDir}`).forNodeFs();
-
-        this.debugLog(`Scanning worktrees directory: ${worktreesDir}`);
-
-        if (!fs.existsSync(worktreesDir)) {
-            this.debugLog(`Worktrees directory does not exist`);
-            return agents;
-        }
-
-        try {
-            const entries = fs.readdirSync(worktreesDir, { withFileTypes: true });
-
-            for (const entry of entries) {
-                if (!entry.isDirectory()) {
-                    continue;
-                }
-
-                // Only look at directories that look like agent worktrees (claude-* or agent-*)
-                if (!entry.name.startsWith('claude-') && !entry.name.startsWith('agent-')) {
-                    continue;
-                }
-
-                const worktreePath = `${repoTerminalPath}/${this.worktreeDir}/${entry.name}`;
-                const metadata = this.loadAgentFromWorktree(worktreePath);
-
-                if (metadata) {
-                    // Update worktreePath in case it was moved
-                    metadata.worktreePath = worktreePath;
-                    metadata.repoPath = repoPath;
-                    agents.push(metadata);
-                    this.debugLog(`Found agent in worktree: ${entry.name}`);
-                }
-            }
-        } catch (error) {
-            this.debugLog(`Failed to scan worktrees: ${error}`);
-        }
-
-        return agents;
-    }
-
-    /**
-     * Check if a worktree exists at the given path
-     */
-    private worktreeExists(worktreePath: string): boolean {
-        const fsPath = agentPath(worktreePath).forNodeFs();
-        return fs.existsSync(fsPath);
-    }
-
-    // Copy coordination files (slash commands, etc.) to a worktree
-    private copyCoordinationToWorktree(agent: Agent): void {
-        const config = getConfigService();
-        const coordinationPath = config.coordinationScriptsPath;
-
-        try {
-            // Use AgentPath for consistent path handling
-            const worktreePath = agentPath(agent.worktreePath);
-            const repoPath = agentPath(agent.repoPath);
-
-            // Create directories in worktree (use forNodeFs() for fs operations)
-            const worktreeCommandsDir = worktreePath.join('.claude', 'commands').forNodeFs();
-            const worktreeAgentsDir = worktreePath.join('.claude-agents').forNodeFs();
-            fs.mkdirSync(worktreeCommandsDir, { recursive: true });
-            fs.mkdirSync(worktreeAgentsDir, { recursive: true });
-
-            // Determine coordination source: configured path, or bundled with extension
-            const bundledCoordPath = agentPath(this.extensionPath).join('coordination');
-            const effectiveCoordPath = coordinationPath
-                ? agentPath(coordinationPath)
-                : bundledCoordPath;
-
-            // Copy slash commands (e.g., claim.md)
-            const commandsSrcDir = effectiveCoordPath.join('commands').forNodeFs();
-            if (fs.existsSync(commandsSrcDir)) {
-                const commands = fs.readdirSync(commandsSrcDir);
-                for (const cmd of commands) {
-                    fs.copyFileSync(
-                        `${commandsSrcDir}/${cmd}`,
-                        `${worktreeCommandsDir}/${cmd}`
-                    );
-                }
-            }
-
-            // Copy task-claimer.sh
-            const claimerSrc = effectiveCoordPath.join('task-claimer.sh').forNodeFs();
-            if (fs.existsSync(claimerSrc)) {
-                const dest = `${worktreeAgentsDir}/task-claimer.sh`;
-                fs.copyFileSync(claimerSrc, dest);
-                try { fs.chmodSync(dest, 0o755); } catch { /* ignore */ }
-            }
-
-            // Copy agent-CLAUDE.md as CLAUDE.md if it exists
-            const claudeMdSrc = effectiveCoordPath.join('agent-CLAUDE.md').forNodeFs();
-            if (fs.existsSync(claudeMdSrc)) {
-                fs.copyFileSync(claudeMdSrc, `${worktreeAgentsDir}/CLAUDE.md`);
-            }
-
-            // Copy hooks for status tracking
-            const hooksSrcDir = effectiveCoordPath.join('hooks').forNodeFs();
-            if (fs.existsSync(hooksSrcDir)) {
-                const worktreeHooksDir = `${worktreeAgentsDir}/hooks`;
-                fs.mkdirSync(worktreeHooksDir, { recursive: true });
-                const hooks = fs.readdirSync(hooksSrcDir);
-                for (const hook of hooks) {
-                    const src = `${hooksSrcDir}/${hook}`;
-                    const dest = `${worktreeHooksDir}/${hook}`;
-                    fs.copyFileSync(src, dest);
-                    try { fs.chmodSync(dest, 0o755); } catch { /* ignore */ }
-                }
-            }
-
-            // Copy hooks.json to .claude/settings.json for this worktree
-            const hooksJsonSrc = effectiveCoordPath.join('hooks.json').forNodeFs();
-            if (fs.existsSync(hooksJsonSrc)) {
-                const worktreeClaudeDir = worktreePath.join('.claude').forNodeFs();
-                fs.mkdirSync(worktreeClaudeDir, { recursive: true });
-                fs.copyFileSync(hooksJsonSrc, `${worktreeClaudeDir}/settings.json`);
-            }
-
-            // Create status directory
-            fs.mkdirSync(`${worktreeAgentsDir}/status`, { recursive: true });
-
-            // Also copy from main repo's .claude-agents if it exists (overrides bundled)
-            const repoAgentsDir = repoPath.join('.claude-agents').forNodeFs();
-            if (fs.existsSync(repoAgentsDir)) {
-                // Copy task-claimer.sh
-                const repoClaimerSrc = `${repoAgentsDir}/task-claimer.sh`;
-                if (fs.existsSync(repoClaimerSrc)) {
-                    const dest = `${worktreeAgentsDir}/task-claimer.sh`;
-                    fs.copyFileSync(repoClaimerSrc, dest);
-                    try { fs.chmodSync(dest, 0o755); } catch { /* ignore */ }
-                }
-
-                // Copy CLAUDE.md
-                const repoClaudeMdSrc = `${repoAgentsDir}/CLAUDE.md`;
-                if (fs.existsSync(repoClaudeMdSrc)) {
-                    fs.copyFileSync(repoClaudeMdSrc, `${worktreeAgentsDir}/CLAUDE.md`);
-                }
-            }
-
-            // Symlink or copy the backlog directory so agents can access tasks
-            const backlogPathSetting = config.backlogPath;
-            if (backlogPathSetting) {
-                const backlogPathObj = agentPath(backlogPathSetting);
-                const worktreeBacklogDir = `${worktreeAgentsDir}/backlog`;
-
-                // Remove existing backlog dir/link
-                try { fs.unlinkSync(worktreeBacklogDir); } catch { /* ignore */ }
-                try { fs.rmdirSync(worktreeBacklogDir, { recursive: true }); } catch { /* ignore */ }
-
-                // Create symlink (junction on Windows)
-                try {
-                    fs.symlinkSync(backlogPathObj.forNodeFs(), worktreeBacklogDir, 'junction');
-                } catch (e) {
-                    // Fallback: copy the directory if symlink fails
-                    // Symlink failed, copy directory instead
-                    this.copyDirRecursive(backlogPathObj.forNodeFs(), worktreeBacklogDir);
-                }
-            }
-
-            // Ensure .opus-orchestra is in .gitignore so agent metadata isn't committed
-            const gitignorePath = worktreePath.join('.gitignore').forNodeFs();
-            let gitignoreContent = '';
-            if (fs.existsSync(gitignorePath)) {
-                gitignoreContent = fs.readFileSync(gitignorePath, 'utf-8');
-            }
-            if (!gitignoreContent.includes('.opus-orchestra')) {
-                const newLine = gitignoreContent.length > 0 && !gitignoreContent.endsWith('\n') ? '\n' : '';
-                fs.writeFileSync(gitignorePath, gitignoreContent + newLine + '.opus-orchestra/\n');
-            }
-        } catch (error) {
-            console.error('[Claude Agents] Failed to copy coordination files:', error);
-        }
-    }
 
     async createAgents(count: number, repoPath?: string, isolationTier?: IsolationTier): Promise<void> {
         const repoPaths = this.getRepositoryPaths();
@@ -527,7 +123,6 @@ export class AgentManager {
             return;
         }
 
-        // If multiple repos configured and none specified, ask user to pick
         let targetRepo = repoPath;
         if (!targetRepo) {
             if (repoPaths.length === 1) {
@@ -544,7 +139,6 @@ export class AgentManager {
             }
         }
 
-        // Check if it's a git repo
         try {
             this.execCommand('git rev-parse --git-dir', targetRepo);
         } catch {
@@ -552,16 +146,12 @@ export class AgentManager {
             return;
         }
 
-        // Get default isolation tier from settings if not specified
         const defaultTier = isolationTier || getConfigService().isolationTier;
-
-        // Load repo-specific container config
         const repoConfig = this.containerManager.loadRepoConfig(targetRepo);
-
-        // Validate isolation tier is available
         const availableTiers = await this.containerManager.getAvailableTiers();
+
         if (!availableTiers.includes(defaultTier)) {
-            const fallback = availableTiers[availableTiers.length - 1];  // Highest available
+            const fallback = availableTiers[availableTiers.length - 1];
             const useAlternative = await vscode.window.showWarningMessage(
                 `Isolation tier '${defaultTier}' is not available. Use '${fallback}' instead?`,
                 'Yes', 'No'
@@ -572,9 +162,7 @@ export class AgentManager {
         }
 
         const baseBranch = this.execCommand('git branch --show-current', targetRepo).trim();
-        const repoTerminalPath = this.toTerminalPath(targetRepo);
 
-        // Find the next available agent ID to avoid duplicates
         const existingIds = new Set<number>();
         for (const agent of this.agents.values()) {
             if (agent.repoPath === targetRepo) {
@@ -584,6 +172,7 @@ export class AgentManager {
 
         let createdCount = 0;
         let restoredCount = 0;
+
         await vscode.window.withProgress({
             location: vscode.ProgressLocation.Notification,
             title: 'Creating agent worktrees',
@@ -591,22 +180,19 @@ export class AgentManager {
         }, async (progress) => {
             let nextId = 1;
             while (createdCount + restoredCount < count) {
-                // Find next available ID
                 while (existingIds.has(nextId)) {
                     nextId++;
                 }
 
-                const agentName = this.getAgentName(nextId);  // Display name (word-based)
+                const agentName = this.getAgentName(nextId);
                 progress.report({ message: `Creating agent ${agentName}...`, increment: (100 / count) });
 
                 const branchName = `claude-${agentName}`;
-                const worktreePath = `${repoTerminalPath}/${this.worktreeDir}/claude-${agentName}`.replace(/\\/g, '/');
+                const worktreePath = this.worktreeManager.getWorktreePath(targetRepo, agentName);
 
                 try {
-                    // Check if worktree already exists - NEVER delete existing worktrees
-                    if (this.worktreeExists(worktreePath)) {
-                        // Try to restore from existing worktree metadata
-                        const existingMetadata = this.loadAgentFromWorktree(worktreePath);
+                    if (this.worktreeManager.worktreeExists(worktreePath)) {
+                        const existingMetadata = this.worktreeManager.loadAgentMetadata(worktreePath);
 
                         if (existingMetadata) {
                             this.debugLog(`Restoring existing agent from worktree: ${agentName}`);
@@ -614,7 +200,7 @@ export class AgentManager {
                             const agent: Agent = {
                                 id: existingMetadata.id || nextId,
                                 name: existingMetadata.name || agentName,
-                                sessionId: existingMetadata.sessionId || this.generateSessionId(),
+                                sessionId: existingMetadata.sessionId || this.persistence.generateSessionId(),
                                 branch: existingMetadata.branch || branchName,
                                 worktreePath,
                                 repoPath: targetRepo,
@@ -633,28 +219,20 @@ export class AgentManager {
                             existingIds.add(agent.id);
                             restoredCount++;
 
-                            // Create terminal for this agent
                             this.createTerminalForAgent(agent, agent.sessionStarted);
-
-                            // Emit agent created event
                             getEventBus().emit('agent:created', { agent });
 
                             nextId++;
                             continue;
-                        } else {
-                            // Worktree exists but no metadata - create metadata for it
-                            this.debugLog(`Worktree exists without metadata, adopting: ${agentName}`);
                         }
                     } else {
-                        // Create new worktree (don't delete branch - it might have commits)
-                        this.execCommand(`git worktree add -B "${branchName}" "${worktreePath}" "${baseBranch}"`, targetRepo);
+                        this.worktreeManager.createWorktree(targetRepo, worktreePath, branchName, baseBranch);
                     }
 
-                    // Create agent entry with new session ID
                     const agent: Agent = {
                         id: nextId,
-                        name: agentName,  // Display name without "claude-" prefix
-                        sessionId: this.generateSessionId(),
+                        name: agentName,
+                        sessionId: this.persistence.generateSessionId(),
                         branch: branchName,
                         worktreePath,
                         repoPath: targetRepo,
@@ -672,13 +250,9 @@ export class AgentManager {
                     existingIds.add(nextId);
                     createdCount++;
 
-                    // Copy coordination files (slash commands, etc.) to the worktree
-                    this.copyCoordinationToWorktree(agent);
+                    this.worktreeManager.copyCoordinationFiles(agent);
+                    this.worktreeManager.saveAgentMetadata(agent);
 
-                    // Save agent metadata to worktree for persistence
-                    this.saveAgentToWorktree(agent);
-
-                    // Create container/sandbox if not standard mode
                     if (defaultTier !== 'standard') {
                         try {
                             const containerInfo = await this.containerManager.createContainer(
@@ -696,10 +270,7 @@ export class AgentManager {
                         }
                     }
 
-                    // Create terminal for this agent
                     this.createTerminalForAgent(agent);
-
-                    // Emit agent created event
                     getEventBus().emit('agent:created', { agent });
 
                 } catch (error) {
@@ -710,8 +281,7 @@ export class AgentManager {
             }
         });
 
-        // Persist agents to storage (both VS Code state and worktree metadata)
-        this.saveAgents();
+        this.persistence.saveAgents(this.agents);
 
         const tierInfo = defaultTier !== 'standard' ? ` (${defaultTier} isolation)` : '';
         const message = restoredCount > 0
@@ -719,6 +289,174 @@ export class AgentManager {
             : `Created ${createdCount} agent worktrees${tierInfo}`;
         vscode.window.showInformationMessage(message);
     }
+
+    async deleteAgent(agentId: number): Promise<boolean> {
+        const agent = this.agents.get(agentId);
+        if (!agent) {
+            return false;
+        }
+
+        if (agent.terminal) {
+            agent.terminal.dispose();
+        }
+
+        if (agent.containerInfo) {
+            await this.containerManager.removeContainer(agentId);
+        }
+
+        try {
+            this.worktreeManager.removeWorktree(agent.repoPath, agent.worktreePath, agent.branch);
+        } catch {
+            // Ignore errors during cleanup
+        }
+
+        this.agents.delete(agentId);
+        this.persistence.saveAgents(this.agents);
+        getEventBus().emit('agent:deleted', { agentId });
+        vscode.window.showInformationMessage(`Agent ${agent.name || agentId} deleted`);
+        return true;
+    }
+
+    async renameAgent(agentId: number, newName: string): Promise<boolean> {
+        const agent = this.agents.get(agentId);
+        if (!agent) {
+            return false;
+        }
+
+        const sanitizedName = newName.trim()
+            .toLowerCase()
+            .replace(/[^a-z0-9-]/g, '-')
+            .replace(/-+/g, '-')
+            .replace(/^-|-$/g, '');
+
+        if (!sanitizedName) {
+            vscode.window.showErrorMessage('Invalid agent name');
+            return false;
+        }
+
+        for (const a of this.agents.values()) {
+            if (a.id !== agentId && a.name === sanitizedName) {
+                vscode.window.showErrorMessage(`Agent name "${sanitizedName}" already exists`);
+                return false;
+            }
+        }
+
+        const oldBranch = agent.branch;
+        const newBranch = `claude-${sanitizedName}`;
+        const newWorktreePath = this.worktreeManager.getWorktreePath(agent.repoPath, sanitizedName);
+        const previousName = agent.name;
+
+        try {
+            if (agent.terminal) {
+                agent.terminal.dispose();
+                agent.terminal = null;
+            }
+
+            this.worktreeManager.renameWorktree(
+                agent.repoPath,
+                agent.worktreePath,
+                newWorktreePath,
+                oldBranch,
+                newBranch
+            );
+
+            agent.name = sanitizedName;
+            agent.branch = newBranch;
+            agent.worktreePath = newWorktreePath;
+
+            this.worktreeManager.copyCoordinationFiles(agent);
+            this.persistence.saveAgents(this.agents);
+
+            getEventBus().emit('agent:renamed', { agent, previousName });
+            vscode.window.showInformationMessage(`Agent renamed to "${sanitizedName}"`);
+            return true;
+        } catch (error) {
+            vscode.window.showErrorMessage(`Failed to rename agent: ${error}`);
+            return false;
+        }
+    }
+
+    async cleanup(): Promise<void> {
+        for (const agent of this.agents.values()) {
+            if (agent.terminal) {
+                agent.terminal.dispose();
+            }
+            if (agent.containerInfo) {
+                await this.containerManager.removeContainer(agent.id);
+            }
+        }
+
+        await vscode.window.withProgress({
+            location: vscode.ProgressLocation.Notification,
+            title: 'Cleaning up worktrees',
+            cancellable: false
+        }, async (progress) => {
+            for (const agent of this.agents.values()) {
+                progress.report({ message: `Removing agent ${agent.id}...` });
+                this.worktreeManager.removeWorktree(agent.repoPath, agent.worktreePath, agent.branch);
+            }
+        });
+
+        this.agents.clear();
+        this.persistence.saveAgents(this.agents);
+        vscode.window.showInformationMessage('Agent worktrees cleaned up');
+    }
+
+    // ========================================================================
+    // Isolation Tier Management
+    // ========================================================================
+
+    async changeAgentIsolationTier(agentId: number, newTier: IsolationTier): Promise<boolean> {
+        const agent = this.agents.get(agentId);
+        if (!agent) {
+            this.debugLog(`Cannot change tier: agent ${agentId} not found`);
+            return false;
+        }
+
+        const currentTier = agent.isolationTier || 'standard';
+        if (currentTier === newTier) {
+            this.debugLog(`Agent ${agentId} already at tier ${newTier}`);
+            return true;
+        }
+
+        this.debugLog(`Changing agent ${agentId} isolation from ${currentTier} to ${newTier}`);
+
+        try {
+            if (currentTier !== 'standard') {
+                await this.containerManager.removeContainer(agentId);
+            }
+
+            if (newTier !== 'standard') {
+                const repoConfig = this.containerManager.loadRepoConfig(agent.repoPath);
+                const containerInfo = await this.containerManager.createContainer(
+                    agentId,
+                    agent.worktreePath,
+                    newTier,
+                    repoConfig
+                );
+                agent.containerInfo = containerInfo;
+            } else {
+                agent.containerInfo = undefined;
+            }
+
+            agent.isolationTier = newTier;
+            this.persistence.saveAgents(this.agents);
+
+            return true;
+        } catch (error) {
+            this.debugLog(`Failed to change isolation tier: ${error}`);
+            vscode.window.showErrorMessage(`Failed to change isolation tier: ${error}`);
+            return false;
+        }
+    }
+
+    async getAgentContainerStats(agentId: number): Promise<{ memoryMB: number; cpuPercent: number } | null> {
+        return this.containerManager.getContainerStats(agentId);
+    }
+
+    // ========================================================================
+    // Terminal Management
+    // ========================================================================
 
     private createTerminalForAgent(agent: Agent, resumeSession: boolean = false): void {
         const config = getConfigService();
@@ -740,53 +478,20 @@ export class AgentManager {
         agent.terminal = terminal;
     }
 
-    async startClaudeInAgent(agentId: number): Promise<void> {
-        const agent = this.agents.get(agentId);
-        if (!agent) {
-            return;
-        }
-
-        // Ensure terminal exists (but don't auto-start Claude)
-        this.ensureTerminalExists(agent);
-
-        const claudeCmd = getConfigService().claudeCommand;
-
-        // Use -r to resume if session was previously started, otherwise create new with --session-id
-        const cmd = agent.sessionStarted
-            ? `${claudeCmd} -r "${agent.sessionId}"`
-            : `${claudeCmd} --session-id "${agent.sessionId}"`;
-        agent.terminal!.show();
-        agent.terminal!.sendText(cmd);
-
-        // Mark session as started so future calls will resume
-        if (!agent.sessionStarted) {
-            agent.sessionStarted = true;
-            this.saveAgents();
-        }
-
-        // Set initial status to waiting-input (Claude starts waiting for user input)
-        agent.status = 'waiting-input';
-        this.updateAgentIcon(agent);
-    }
-
-    // Ensure agent has a terminal (without starting Claude)
     private ensureTerminalExists(agent: Agent): void {
         const terminalService = getTerminalService();
 
-        // Check if we have a valid terminal that still exists
         if (agent.terminal && terminalService.isTerminalAlive(agent.terminal)) {
             return;
         }
         agent.terminal = null;
 
-        // Try to find existing terminal by name
         const existingTerminal = terminalService.findTerminalByName(agent.name);
         if (existingTerminal) {
             agent.terminal = existingTerminal;
             return;
         }
 
-        // Create terminal without starting Claude
         agent.terminal = terminalService.createTerminal({
             name: agent.name,
             cwd: agent.worktreePath,
@@ -794,12 +499,29 @@ export class AgentManager {
         });
     }
 
-    getAgents(): Agent[] {
-        return Array.from(this.agents.values());
-    }
+    async startClaudeInAgent(agentId: number): Promise<void> {
+        const agent = this.agents.get(agentId);
+        if (!agent) {
+            return;
+        }
 
-    getAgent(id: number): Agent | undefined {
-        return this.agents.get(id);
+        this.ensureTerminalExists(agent);
+
+        const claudeCmd = getConfigService().claudeCommand;
+        const cmd = agent.sessionStarted
+            ? `${claudeCmd} -r "${agent.sessionId}"`
+            : `${claudeCmd} --session-id "${agent.sessionId}"`;
+
+        agent.terminal!.show();
+        agent.terminal!.sendText(cmd);
+
+        if (!agent.sessionStarted) {
+            agent.sessionStarted = true;
+            this.persistence.saveAgents(this.agents);
+        }
+
+        agent.status = 'waiting-input';
+        this.statusTracker.updateAgentIcon(agent);
     }
 
     focusAgent(agentId: number): void {
@@ -809,14 +531,12 @@ export class AgentManager {
             return;
         }
 
-        // Ensure we have a terminal (without auto-starting Claude)
         this.ensureTerminalExists(agent);
 
         if (agent.terminal) {
-            agent.terminal.show(true); // true = preserve focus
+            agent.terminal.show(true);
         }
     }
-
 
     sendToAgent(agentId: number, text: string): void {
         const agent = this.agents.get(agentId);
@@ -826,7 +546,7 @@ export class AgentManager {
             agent.pendingApproval = null;
             agent.status = 'working';
             agent.lastInteractionTime = new Date();
-            this.updateAgentIcon(agent);
+            this.statusTracker.updateAgentIcon(agent);
 
             if (hadPendingApproval) {
                 getEventBus().emit('approval:resolved', { agentId });
@@ -839,250 +559,44 @@ export class AgentManager {
             if (agent.terminal === terminal) {
                 agent.terminal = null;
                 agent.status = 'idle';
-                this.updateAgentIcon(agent);
+                this.statusTracker.updateAgentIcon(agent);
                 getEventBus().emit('agent:terminalClosed', { agentId: agent.id });
                 break;
             }
         }
     }
 
+    // ========================================================================
+    // Status & Queries (delegated to StatusTracker)
+    // ========================================================================
+
     refreshStatus(): void {
-        this.debugLog(`[refreshStatus] called, agents count: ${this.agents.size}`);
-        for (const agent of this.agents.values()) {
-            // Check hook-based status file
-            this.checkHookStatus(agent);
-            this.updateAgentIcon(agent);
-        }
+        this.statusTracker.refreshStatus(this.agents);
     }
 
-    // Read status from hook-generated files using StatusService
-    private checkHookStatus(agent: Agent): void {
-        const parsedStatus = getStatusService().checkStatus(agent.worktreePath);
-        if (parsedStatus) {
-            const previousStatus = agent.status;
-            const hadApproval = agent.pendingApproval !== null;
-
-            agent.status = parsedStatus.status;
-            agent.pendingApproval = parsedStatus.pendingApproval;
-
-            // Emit status change event if status actually changed
-            if (previousStatus !== agent.status) {
-                getEventBus().emit('agent:statusChanged', { agent, previousStatus });
-            }
-
-            // Emit approval pending event if new approval appeared
-            if (!hadApproval && agent.pendingApproval !== null) {
-                getEventBus().emit('approval:pending', {
-                    approval: {
-                        agentId: agent.id,
-                        description: agent.pendingApproval,
-                        timestamp: new Date(),
-                    }
-                });
-            }
-        }
-    }
-
-    // Separate async diff refresh - call this on a longer interval
     async refreshDiffStats(): Promise<void> {
-        const diffInterval = getConfigService().diffPollingInterval;
-
-        // If disabled, skip
-        if (diffInterval === 0) {
-            return;
-        }
-
-        // Run all diff commands in parallel using async exec
-        const promises: Promise<void>[] = [];
-
-        for (const agent of this.agents.values()) {
-            promises.push(this.getDiffStatsAsync(agent));
-        }
-
-        await Promise.all(promises);
-    }
-
-    private async getDiffStatsAsync(agent: Agent): Promise<void> {
-        try {
-            const gitService = getGitService();
-            const baseBranch = await gitService.getBaseBranch(agent.repoPath);
-            agent.diffStats = await gitService.getDiffStats(agent.worktreePath, baseBranch);
-        } catch {
-            // Keep existing stats on error
-        }
-    }
-
-    private async getBaseBranchAsync(repoPath?: string): Promise<string> {
-        const cwd = repoPath || this.workspaceRoot;
-        return getGitService().getBaseBranch(cwd);
-    }
-
-    private execCommandAsync(command: string, cwd: string): Promise<string> {
-        return getCommandService().execAsync(command, cwd);
-    }
-
-
-    private updateAgentIcon(agent: Agent): void {
-        if (agent.status === 'idle') {
-            // Idle status depends on terminal state
-            agent.statusIcon = agent.terminal ? 'circle-filled' : 'circle-outline';
-        } else {
-            agent.statusIcon = STATUS_ICONS[agent.status];
-        }
+        await this.statusTracker.refreshDiffStats(this.agents);
     }
 
     getPendingApprovals(): PendingApproval[] {
-        const approvals: PendingApproval[] = [];
-        for (const agent of this.agents.values()) {
-            if (agent.pendingApproval) {
-                approvals.push({
-                    agentId: agent.id,
-                    description: agent.pendingApproval,
-                    timestamp: new Date()
-                });
-            }
-        }
-        return approvals;
+        return this.statusTracker.getPendingApprovals(this.agents);
     }
 
     getWaitingCount(): number {
-        let count = 0;
-        for (const agent of this.agents.values()) {
-            if (agent.status === 'waiting-input' || agent.status === 'waiting-approval') {
-                count++;
-            }
-        }
-        return count;
+        return this.statusTracker.getWaitingCount(this.agents);
     }
 
-    async cleanup(): Promise<void> {
-        // Close all terminals and containers
-        for (const agent of this.agents.values()) {
-            if (agent.terminal) {
-                agent.terminal.dispose();
-            }
-            // Remove container if exists
-            if (agent.containerInfo) {
-                await this.containerManager.removeContainer(agent.id);
-            }
-        }
-
-        // Remove worktrees
-        await vscode.window.withProgress({
-            location: vscode.ProgressLocation.Notification,
-            title: 'Cleaning up worktrees',
-            cancellable: false
-        }, async (progress) => {
-            for (const agent of this.agents.values()) {
-                progress.report({ message: `Removing agent ${agent.id}...` });
-                this.execCommandSilent(`git worktree remove "${agent.worktreePath}" --force`, agent.repoPath);
-                this.execCommandSilent(`git branch -D "${agent.branch}"`, agent.repoPath);
-            }
-        });
-
-        this.agents.clear();
-        this.saveAgents();
-        vscode.window.showInformationMessage('Agent worktrees cleaned up');
+    getAgents(): Agent[] {
+        return Array.from(this.agents.values());
     }
 
-    // Delete a single agent
-    async deleteAgent(agentId: number): Promise<boolean> {
-        const agent = this.agents.get(agentId);
-        if (!agent) {
-            return false;
-        }
-
-        // Close terminal if open
-        if (agent.terminal) {
-            agent.terminal.dispose();
-        }
-
-        // Remove container if exists
-        if (agent.containerInfo) {
-            await this.containerManager.removeContainer(agentId);
-        }
-
-        // Remove worktree and branch
-        try {
-            this.execCommandSilent(`git worktree remove "${agent.worktreePath}" --force`, agent.repoPath);
-            this.execCommandSilent(`git branch -D "${agent.branch}"`, agent.repoPath);
-        } catch {
-            // Ignore errors during cleanup
-        }
-
-        this.agents.delete(agentId);
-        this.saveAgents();
-        getEventBus().emit('agent:deleted', { agentId });
-        vscode.window.showInformationMessage(`Agent ${agent.name || agentId} deleted`);
-        return true;
+    getAgent(id: number): Agent | undefined {
+        return this.agents.get(id);
     }
 
-    async renameAgent(agentId: number, newName: string): Promise<boolean> {
-        const agent = this.agents.get(agentId);
-        if (!agent) {
-            return false;
-        }
-
-        // Sanitize the name (remove special characters, spaces -> hyphens)
-        const sanitizedName = newName.trim()
-            .toLowerCase()
-            .replace(/[^a-z0-9-]/g, '-')
-            .replace(/-+/g, '-')
-            .replace(/^-|-$/g, '');
-
-        if (!sanitizedName) {
-            vscode.window.showErrorMessage('Invalid agent name');
-            return false;
-        }
-
-        const oldBranch = agent.branch;
-        const newBranch = `claude-${sanitizedName}`;
-        const terminalPath = this.toTerminalPath(agent.repoPath);
-        const newWorktreePath = `${terminalPath}/${this.worktreeDir}/claude-${sanitizedName}`.replace(/\\/g, '/');
-
-        // Check if name already exists
-        for (const a of this.agents.values()) {
-            if (a.id !== agentId && a.name === sanitizedName) {
-                vscode.window.showErrorMessage(`Agent name "${sanitizedName}" already exists`);
-                return false;
-            }
-        }
-
-        const previousName = agent.name;
-
-        try {
-            // Close terminal if open
-            if (agent.terminal) {
-                agent.terminal.dispose();
-                agent.terminal = null;
-            }
-
-            // Move worktree: remove old, create new with same branch content
-            this.execCommandSilent(`git worktree remove "${agent.worktreePath}" --force`, agent.repoPath);
-
-            // Rename branch
-            this.execCommand(`git branch -m "${oldBranch}" "${newBranch}"`, agent.repoPath);
-
-            // Create new worktree
-            this.execCommand(`git worktree add "${newWorktreePath}" "${newBranch}"`, agent.repoPath);
-
-            // Update agent
-            agent.name = sanitizedName;
-            agent.branch = newBranch;
-            agent.worktreePath = newWorktreePath;
-
-            // Re-copy coordination files to new worktree
-            this.copyCoordinationToWorktree(agent);
-
-            this.saveAgents();
-            getEventBus().emit('agent:renamed', { agent, previousName });
-            vscode.window.showInformationMessage(`Agent renamed to "${sanitizedName}"`);
-            return true;
-        } catch (error) {
-            vscode.window.showErrorMessage(`Failed to rename agent: ${error}`);
-            return false;
-        }
-    }
+    // ========================================================================
+    // Git & Project Operations
+    // ========================================================================
 
     async showAgentDiff(agentId: number): Promise<void> {
         const agent = this.agents.get(agentId);
@@ -1091,11 +605,10 @@ export class AgentManager {
         }
 
         try {
-            const baseBranch = await this.getBaseBranchAsync(agent.repoPath);
+            const baseBranch = await getGitService().getBaseBranch(agent.repoPath);
             const windowsWorktreePath = this.toWindowsPath(agent.worktreePath);
 
-            // Get list of changed files
-            const output = await this.execCommandAsync(
+            const output = await getCommandService().execAsync(
                 `git diff --name-only ${baseBranch}...HEAD`,
                 agent.worktreePath
             );
@@ -1107,8 +620,6 @@ export class AgentManager {
                 return;
             }
 
-            // Open VS Code's SCM diff view for the worktree
-            // Use git.openChange command for each file, or show quick pick
             const items = changedFiles.map(f => ({
                 label: f,
                 description: `${agent.name}: ${f}`
@@ -1123,11 +634,9 @@ export class AgentManager {
                 const filePath = path.join(windowsWorktreePath, selected.label);
                 const uri = vscode.Uri.file(filePath);
 
-                // Try to use git extension's diff view
                 try {
                     await vscode.commands.executeCommand('git.openChange', uri);
                 } catch {
-                    // Fallback: just open the file
                     await vscode.commands.executeCommand('vscode.open', uri);
                 }
             }
@@ -1157,9 +666,7 @@ export class AgentManager {
             return { success: false, message: 'No workspace folder open' };
         }
 
-        // Check if already a git repo
         if (this.isGitRepo()) {
-            // Add .worktrees to .gitignore if not present
             const gitignorePath = path.join(this.workspaceRoot, '.gitignore');
             let gitignoreContent = '';
 
@@ -1175,11 +682,9 @@ export class AgentManager {
             return { success: true, message: 'Project already initialized. Added .worktrees to .gitignore.' };
         }
 
-        // Initialize git repo
         try {
             this.execCommand('git init', this.workspaceRoot);
 
-            // Create .gitignore with .worktrees
             const gitignorePath = path.join(this.workspaceRoot, '.gitignore');
             let gitignoreContent = '';
 
@@ -1192,12 +697,11 @@ export class AgentManager {
                 fs.writeFileSync(gitignorePath, gitignoreContent + newLine + '.worktrees/\n');
             }
 
-            // Initial commit so worktrees can be created
             this.execCommand('git add -A', this.workspaceRoot);
             try {
                 this.execCommand('git commit -m "Initial commit"', this.workspaceRoot);
             } catch {
-                // Might fail if nothing to commit, that's ok
+                // Might fail if nothing to commit
             }
 
             return { success: true, message: 'Git repository initialized and configured for Claude Agents.' };
@@ -1206,15 +710,16 @@ export class AgentManager {
         }
     }
 
-    // Get available tasks from backlog
+    // ========================================================================
+    // Task Management
+    // ========================================================================
+
     getAvailableTasks(repoPath: string): string[] {
         const terminalPath = this.toTerminalPath(repoPath);
         const backlogPath = `${terminalPath}/.claude-agents/backlog`.replace(/\\/g, '/');
 
         try {
-            // Convert to Windows path for fs operations
             const fsPath = agentPath(backlogPath).forNodeFs();
-
             const files = fs.readdirSync(fsPath);
             return files
                 .filter(f => f.endsWith('.md'))
@@ -1224,7 +729,6 @@ export class AgentManager {
         }
     }
 
-    // Create an agent for a specific task
     async createAgentForTask(taskName: string, repoPath?: string): Promise<Agent | null> {
         const repoPaths = this.getRepositoryPaths();
 
@@ -1249,7 +753,6 @@ export class AgentManager {
             }
         }
 
-        // Check if it's a git repo
         try {
             this.execCommand('git rev-parse --git-dir', targetRepo);
         } catch {
@@ -1260,25 +763,20 @@ export class AgentManager {
         const baseBranch = this.execCommand('git branch --show-current', targetRepo).trim();
         const repoTerminalPath = this.toTerminalPath(targetRepo);
 
-        // Create branch/worktree named after task
         const branchName = `agent-${taskName}`;
         const worktreePath = `${repoTerminalPath}/${this.worktreeDir}/agent-${taskName}`.replace(/\\/g, '/');
         const agentId = this.agents.size + 1;
 
         try {
-            // Remove existing worktree if present
             this.execCommandSilent(`git worktree remove "${worktreePath}" --force`, targetRepo);
-
-            // Delete branch if exists
             this.execCommandSilent(`git branch -D "${branchName}"`, targetRepo);
 
-            // Create worktree with new branch
             this.execCommand(`git worktree add -b "${branchName}" "${worktreePath}" "${baseBranch}"`, targetRepo);
 
             const agent: Agent = {
                 id: agentId,
                 name: taskName,
-                sessionId: this.generateSessionId(),
+                sessionId: this.persistence.generateSessionId(),
                 branch: branchName,
                 worktreePath,
                 repoPath: targetRepo,
@@ -1292,7 +790,7 @@ export class AgentManager {
             };
 
             this.agents.set(agentId, agent);
-            this.saveAgents();
+            this.persistence.saveAgents(this.agents);
             this.createTerminalForAgent(agent);
             getEventBus().emit('agent:created', { agent });
 
@@ -1305,21 +803,17 @@ export class AgentManager {
         }
     }
 
-    // Initialize coordination files in a repository
     async initializeCoordination(repoPath: string, backlogPath?: string): Promise<{ success: boolean; message: string }> {
-        // Use Windows paths for fs operations
         const windowsRepoPath = this.toWindowsPath(repoPath);
         const agentsDir = path.join(windowsRepoPath, '.claude-agents');
         const claudeCommandsDir = path.join(windowsRepoPath, '.claude', 'commands');
         const claudeSkillsDir = path.join(windowsRepoPath, '.claude', 'skills');
 
         try {
-            // Create directories
             fs.mkdirSync(path.join(agentsDir, 'completed'), { recursive: true });
             fs.mkdirSync(claudeCommandsDir, { recursive: true });
             fs.mkdirSync(claudeSkillsDir, { recursive: true });
 
-            // Copy coordination script
             const coordinationPath = getConfigService().coordinationScriptsPath;
 
             if (coordinationPath) {
@@ -1329,19 +823,15 @@ export class AgentManager {
 
                 if (fs.existsSync(scriptSrc)) {
                     fs.copyFileSync(scriptSrc, scriptDest);
-                    try {
-                        fs.chmodSync(scriptDest, 0o755);
-                    } catch { /* ignore on Windows */ }
+                    try { fs.chmodSync(scriptDest, 0o755); } catch { /* ignore */ }
                 }
 
-                // Copy CLAUDE.md
                 const claudeMdSrc = path.join(windowsCoordPath, 'agent-CLAUDE.md');
                 const claudeMdDest = path.join(agentsDir, 'CLAUDE.md');
                 if (fs.existsSync(claudeMdSrc)) {
                     fs.copyFileSync(claudeMdSrc, claudeMdDest);
                 }
 
-                // Copy slash commands
                 const commandsSrcDir = path.join(windowsCoordPath, 'commands');
                 if (fs.existsSync(commandsSrcDir)) {
                     const commands = fs.readdirSync(commandsSrcDir);
@@ -1353,37 +843,27 @@ export class AgentManager {
                     }
                 }
 
-                // Copy skills
                 const skillsSrcDir = path.join(windowsCoordPath, 'skills');
                 if (fs.existsSync(skillsSrcDir)) {
                     this.copyDirRecursive(skillsSrcDir, claudeSkillsDir);
                 }
             }
 
-            // Create or link backlog
             const backlogDir = path.join(agentsDir, 'backlog');
             if (backlogPath) {
                 const windowsBacklogPath = this.toWindowsPath(backlogPath);
-                // Remove existing
-                try {
-                    fs.unlinkSync(backlogDir);
-                } catch { /* ignore */ }
-                try {
-                    fs.rmdirSync(backlogDir, { recursive: true });
-                } catch { /* ignore */ }
-                // Create symlink (use junction on Windows for directories)
+                try { fs.unlinkSync(backlogDir); } catch { /* ignore */ }
+                try { fs.rmdirSync(backlogDir, { recursive: true }); } catch { /* ignore */ }
                 fs.symlinkSync(windowsBacklogPath, backlogDir, 'junction');
             } else {
                 fs.mkdirSync(backlogDir, { recursive: true });
             }
 
-            // Initialize claims file
             const claimsFile = path.join(agentsDir, 'claims.jsonl');
             if (!fs.existsSync(claimsFile)) {
                 fs.writeFileSync(claimsFile, '');
             }
 
-            // Update .gitignore
             const gitignorePath = path.join(windowsRepoPath, '.gitignore');
             let gitignoreContent = '';
             if (fs.existsSync(gitignorePath)) {
@@ -1409,7 +889,6 @@ export class AgentManager {
         }
     }
 
-    // Clean up completed tasks
     async cleanupCompletedTasks(repoPath?: string): Promise<{ success: boolean; count: number }> {
         const repoPaths = repoPath ? [repoPath] : this.getRepositoryPaths();
         let totalCleaned = 0;
@@ -1430,5 +909,21 @@ export class AgentManager {
         }
 
         return { success: true, count: totalCleaned };
+    }
+
+    private copyDirRecursive(src: string, dest: string): void {
+        fs.mkdirSync(dest, { recursive: true });
+        const entries = fs.readdirSync(src, { withFileTypes: true });
+
+        for (const entry of entries) {
+            const srcPath = path.join(src, entry.name);
+            const destPath = path.join(dest, entry.name);
+
+            if (entry.isDirectory()) {
+                this.copyDirRecursive(srcPath, destPath);
+            } else {
+                fs.copyFileSync(srcPath, destPath);
+            }
+        }
     }
 }
