+++ vscode-extension/src/managers/index.ts
@@ -1,0 +1,3 @@
+export { WorktreeManager } from './WorktreeManager';
+export { AgentStatusTracker } from './AgentStatusTracker';
+export { AgentPersistence } from './AgentPersistence';
